+++ src/Utilities/ShadeRec.h
@@ -26,7 +26,7 @@ class ShadeRec {
 		Normal				normal;				// Normal at hit point
 		Ray					ray;				// Required for specular highlights and area lights
 		int					depth;				// recursion depth
-		float				t;					// ray parameter
+		double				t;					// ray parameter
 		World&				w;					// World reference
 		RGBColor            color;
